+++ src/lib.rs
@@ -1,11 +1,12 @@
-extern crate async_std;
-extern crate async_tls;
 extern crate async_trait;
 extern crate cassandra_proto;
 extern crate futures;
 extern crate log;
 extern crate lz4_compress;
 extern crate snap;
+extern crate transport;
+#[cfg(feature = "async_std")]
+extern crate transport_async;
 
 pub mod authenticators;
 pub mod query;
@@ -15,9 +16,6 @@ pub(crate) mod frame_channel;
 mod compressor;
 mod pager;
 mod session;
-mod transport;
-mod transport_tcp;
-mod transport_tls;
 mod utils;
 
 pub use cassandra_proto::compression::Compressor;
@@ -25,5 +23,5 @@ pub use compressor::Compression;
 pub use pager::PageSize;
 pub use session::Session;
 pub use transport::CDRSTransport;
-pub use transport_tcp::TransportTcp;
-pub use transport_tls::TransportTls;
+pub use transport_async::transport_tcp::TransportTcp;
+pub use transport_async::transport_tls::TransportTls;
