+++ src/utils.rs
@@ -1,15 +1,15 @@
 use cassandra_proto::frame::Flag;
 
 pub fn prepare_flags(with_tracing: bool, with_warnings: bool) -> Vec<Flag> {
-  let mut flags = vec![];
+    let mut flags = vec![];
 
-  if with_tracing {
-    flags.push(Flag::Tracing);
-  }
+    if with_tracing {
+        flags.push(Flag::Tracing);
+    }
 
-  if with_warnings {
-    flags.push(Flag::Warning);
-  }
+    if with_warnings {
+        flags.push(Flag::Warning);
+    }
 
-  flags
+    flags
 }
