+++ tests/utils_session.rs
@@ -1,12 +1,11 @@
 use cdrs_async::{authenticators::NoneAuthenticator, Compression, Session, TransportTcp};
 
 pub async fn connect_tcp() -> Session<TransportTcp> {
-  let authenticator_strategy = NoneAuthenticator {};
-  Session::connect(
-    "127.0.0.1:9042",
-    Compression::None,
-    authenticator_strategy.into(),
-  )
-  .await
-  .expect("session connect")
+    let authenticator_strategy = NoneAuthenticator {};
+    let transport = TransportTcp::new("127.0.0.1:9042")
+        .await
+        .expect("Cannot create transport");
+    Session::connect(transport, Compression::None, authenticator_strategy.into())
+        .await
+        .expect("session connect")
 }
