+++ transport_async/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "transport_async"
+version = "0.1.0"
+authors = ["Alex Pikalov <alex.pikalov.khar@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+transport = { path = "../transport"}
+async-std = "1.5.0"
+async-tls = "0.7"
+async-trait = "0.1.30"
