+++ transport_async/src/transport_tls.rs
@@ -10,37 +10,40 @@ use async_std::{
   io::{Read, Write},
   net,
 };
+use async_tls::{client::TlsStream, TlsConnector};
 use async_trait::async_trait;
 
-use super::transport::CDRSTransport;
+use transport::CDRSTransport;
 
-/// CDRS TCP transport.
-pub struct TransportTcp {
-  tcp: net::TcpStream,
+pub type Stream = TlsStream<net::TcpStream>;
+
+/// CDRS TLS transport.
+pub struct TransportTls {
+  stream: Stream,
   _addr: String,
 }
 
-impl TransportTcp {
+impl TransportTls {
   /// Constructs a new `TransportTcp`.
-  pub async fn new(addr: &str) -> io::Result<TransportTcp> {
-    net::TcpStream::connect(addr)
-      .await
-      .map(|socket| TransportTcp {
-        tcp: socket,
-        _addr: addr.to_string(),
-      })
+  pub async fn new(addr: &str, connector: TlsConnector) -> io::Result<TransportTls> {
+    let tcp_stream = net::TcpStream::connect(addr).await?;
+    let stream = connector.connect(addr, tcp_stream).await?;
+    Ok(TransportTls {
+      stream,
+      _addr: addr.to_string(),
+    })
   }
 }
 
-impl Unpin for TransportTcp {}
+impl Unpin for TransportTls {}
 
-impl Read for TransportTcp {
+impl Read for TransportTls {
   fn poll_read(
     mut self: Pin<&mut Self>,
     cx: &mut Context,
     buf: &mut [u8],
   ) -> Poll<io::Result<usize>> {
-    Pin::new(&mut self.tcp).poll_read(cx, buf)
+    Pin::new(&mut self.stream).poll_read(cx, buf)
   }
 
   fn poll_read_vectored(
@@ -48,17 +51,17 @@ impl Read for TransportTcp {
     cx: &mut Context<'_>,
     bufs: &mut [IoSliceMut<'_>],
   ) -> Poll<io::Result<usize>> {
-    Pin::new(&mut self.tcp).poll_read_vectored(cx, bufs)
+    Pin::new(&mut self.stream).poll_read_vectored(cx, bufs)
   }
 }
 
-impl Write for TransportTcp {
+impl Write for TransportTls {
   fn poll_write(
     mut self: Pin<&mut Self>,
     cx: &mut Context<'_>,
     buf: &[u8],
   ) -> Poll<io::Result<usize>> {
-    Pin::new(&mut self.tcp).poll_write(cx, buf)
+    Pin::new(&mut self.stream).poll_write(cx, buf)
   }
 
   fn poll_write_vectored(
@@ -66,35 +69,31 @@ impl Write for TransportTcp {
     cx: &mut Context<'_>,
     bufs: &[IoSlice<'_>],
   ) -> Poll<io::Result<usize>> {
-    Pin::new(&mut self.tcp).poll_write_vectored(cx, bufs)
+    Pin::new(&mut self.stream).poll_write_vectored(cx, bufs)
   }
 
   fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-    Pin::new(&mut self.tcp).poll_flush(cx)
+    Pin::new(&mut self.stream).poll_flush(cx)
   }
 
   fn poll_close(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-    Pin::new(&mut self.tcp).poll_close(cx)
+    Pin::new(&mut self.stream).poll_close(cx)
   }
 }
 
 #[async_trait]
-impl CDRSTransport for TransportTcp {
+impl CDRSTransport for TransportTls {
   // FIXME:
-  // async fn try_clone(&self) -> io::Result<TransportTcp> {
-  //   net::TcpStream::connect(self.addr.as_str())
-  //     .await
-  //     .map(|socket| TransportTcp {
-  //       tcp: socket,
-  //       addr: self.addr.clone(),
-  //     })
+  // async fn try_clone(&self) -> io::Result<TransportTls> {
+  //   // TODO:
+  //   todo!()
   // }
 
   fn close(&mut self, close: net::Shutdown) -> io::Result<()> {
-    self.tcp.shutdown(close)
+    self.stream.get_mut().shutdown(close)
   }
 
   fn is_alive(&self) -> bool {
-    self.tcp.peer_addr().is_ok()
+    self.stream.get_ref().peer_addr().is_ok()
   }
 }
